+++ frontend/srcs/src/app/components/channel/channel-info/channel-info.component.ts
@@ -1,5 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { Channel } from 'src/app/models';
+import { ChatComponent } from 'src/app/modules';
 
 @Component({
   selector: 'app-channel-info',
@@ -8,4 +9,13 @@ import { Channel } from 'src/app/models';
 })
 export class ChannelInfoComponent {
   @Input() channel!: Channel;
+
+  constructor(
+    private readonly chatComponent: ChatComponent
+  ) 
+  {  }
+
+  manageChannel(){
+    this.chatComponent.manageChannel(this.channel);
+  }
 }