+++ frontend/srcs/src/app/components/channel/index.ts
@@ -3,3 +3,6 @@ export * from './channel-info/channel-info.component'
 export * from './channel-sidebar/channel-sidebar.component'
 export * from './channel-window/channel-window.component'
 export * from './channel-window-message/channel-window-message.component'
+export * from './channel-management/channel-management.component'
+export * from './channel-management-users/channel-management-users.component'
+